+++ src/exceptiongroup/__init__.py
@@ -1,4 +1,10 @@
-__all__ = ["BaseExceptionGroup", "ExceptionGroup", "catch"]
+__all__ = [
+    "BaseExceptionGroup",
+    "ExceptionGroup",
+    "catch",
+    "format_exception",
+    "format_exception_only",
+]
 
 import os
 import sys
@@ -8,6 +14,7 @@ from ._version import version as __version__  # noqa: F401
 
 if sys.version_info < (3, 11):
     from ._exceptions import BaseExceptionGroup, ExceptionGroup
+    from ._formatting import format_exception, format_exception_only
 
     if os.getenv("EXCEPTIONGROUP_NO_PATCH") != "1":
         from . import _formatting  # noqa: F401
@@ -15,5 +22,7 @@ if sys.version_info < (3, 11):
     BaseExceptionGroup.__module__ = __name__
     ExceptionGroup.__module__ = __name__
 else:
+    from traceback import format_exception, format_exception_only
+
     BaseExceptionGroup = BaseExceptionGroup
     ExceptionGroup = ExceptionGroup